Add soft links to Dockerfiles in local copies of submodules.
Ansible playbooks will create images for moshimoji stack from
Dockerfiles stored in moshimoji/docker/Dockerfiles. However,
because the Dockerfile must be stored in the build context, these
Dockerfiles are links to the actual files in my local copies of
the submodules.

I could instead use links to the copies of the Dockerfile stored
in moshimoji, but because I may want to test before making commits,
I've chosen the development repos of the submodules to serve as
the build contexts instead of the latest copies of the submodules
in moshimoji.