Update migrate playbooks to work on prod.
Changes include running the show migration/migrate on each node,
and only executing show migration/migrate if the node is the back-
end node.

Running the migrate related tasks only on the manager may fail if
the swarm is bigger than one node (as the prod swarm is) since the
backend task may be on a different node.

It is possible to introspect which node hosts the backend task
from the manager and then ssh from the manager to that node, but
that requires provisioning the nodes with ssh keys and so on.