Update db-sync playbook to work with prod setup.
This change runs the psql command to apply the sql on the remote
swarm if the db is for prod and runs the psql command locally if
the db is for stage.

The psql command to apply the sql must be done on the remote
swarm for prod because the RDS instance only allows connections
from within the same VPC. The psql command to apply the sql cannot
be done on the remote swarm for stage (unless it is done from
within a swarm task) because localhost:5432 refuses connection.

I opted to apply the sql using the psql client on the controller
host rather than a swarm task for stage for simplicity's sake.